+++ client/e2e/src/app.e2e-spec.ts
@@ -10,7 +10,7 @@ describe('workspace-project App', () => {
 
   it('should display welcome message', async () => {
     await page.navigateTo();
-    expect(await page.getTitleText()).toEqual('client app is running!');
+    expect(await page.getTitleText()).toEqual('oerc: TV program');
   });
 
   afterEach(async () => {