+++ orf.go
@@ -20,7 +20,6 @@ package main
 import (
 	"fmt"
 	"github.com/PuerkitoBio/goquery"
-	"github.com/alitto/pond"
 	"github.com/gocolly/colly/v2"
 	"log"
 	"regexp"
@@ -45,59 +44,19 @@ var (
 
 // ORFParser struct to group orf parsing code
 type ORFParser struct {
+	ParserInterface
 	Parser
 }
 
-// Fetch central method to parse ORF tv show and program data
-func (o *ORFParser) Fetch() {
-	db := getDb()
-	o.db = db
+func (o *ORFParser) postProcess() {}
 
-	// get channel family db record
-	var channelFamily = getChannelFamily(db, o.ChannelFamilyKey)
-	if channelFamily.ID == 0 {
-		log.Fatalln("ORF channelFamily was not found!")
-		return
-	}
-	o.ChannelFamily = *channelFamily
-
-	// import tv shows
-	if GetAppConf().EnableTVShowCollection {
-		o.fetchTvShows()
-	}
-
-	timeRange := o.dateRangeHandler.getDateRange()
-	var times []time.Time
-	if timeRange != nil {
-		times = *timeRange
-	} else {
-		log.Printf("No valid time range received!\n")
-		return
-	}
-	// import program entries for the configured date range
-	if GetAppConf().EnableProgramEntryCollection {
-		pool := pond.New(4, 100, getWorkerPoolIdleTimeout())
-		for _, channel := range getChannelsOfFamily(db, channelFamily) {
-			for _, day := range times {
-				chn := channel
-				dayToFetch := day
-
-				pool.Submit(func() {
-					o.handleDay(chn, dayToFetch)
-				})
-			}
-		}
-		// wait for finish
-		pool.StopAndWait()
-	}
-
-	if verboseGlobal {
-		log.Println("ORF parsed successfully")
-	}
+func (o *ORFParser) preProcess() bool {
+	o.parallelWorkersCount = 4
+	return true
 }
 
 // fetchTvShow: This method checks all the tv shows
-func (o *ORFParser) fetchTvShows() {
+func (o *ORFParser) fetchTVShows() {
 	if !GetAppConf().EnableTVShowCollection || isRecentlyFetched() {
 		log.Printf("Skip update of tv shows, due to recent fetch. Use 'forceUpdate' = true to ignore this.")
 		return
@@ -150,7 +109,7 @@ func (o *ORFParser) fetchTvShows() {
 
 	err := c.Visit(orfHostWithPrefix + "/profiles")
 	if err != nil {
-		appLog(fmt.Sprintf("ORF parser fetch url error: %v\n", err))
+		appLog(fmt.Sprintf("ORF parser fetch url error: %v", err))
 	}
 	c.Wait()
 }
@@ -280,7 +239,7 @@ func (o *ORFParser) handleDay(channel Channel, day time.Time) {
 			requestHeaders := map[string]string{"Origin": "tv.orf.at", "Host": "tv.orf.at", "Accept": "text/html"}
 			response, err := doGetRequest(url, requestHeaders, 3)
 			if err != nil || response == nil {
-				appLog(fmt.Sprintf("Problem fetching orf URL '%s'\n", url))
+				appLog(fmt.Sprintf("Problem fetching orf URL '%s'", url))
 				return
 			}
 			// Load the HTML document
@@ -325,7 +284,7 @@ func (o *ORFParser) handleDay(channel Channel, day time.Time) {
 
 	err = c.Visit(fmt.Sprintf("%s%s", orfProgramHostWithPrefix, programDetailURLPerDay))
 	if err != nil {
-		appLog(fmt.Sprintf("Error of orf collector in url '%s': %v\n", queryURL, err))
+		appLog(fmt.Sprintf("Error of orf collector in url '%s': %v", queryURL, err))
 	}
 	c.Wait()
 }
@@ -334,6 +293,21 @@ func (o *ORFParser) getDaysBetween(day time.Time, now time.Time) int {
 	first := time.Date(day.Year(), day.Month(), day.Day(), 0, 0, 0, 0, time.UTC)
 	second := time.Date(now.Year(), now.Month(), now.Day(), 0, 0, 0, 0, time.UTC)
 	return int(first.Sub(second).Hours() / 24)
+}
+
+func (o *ORFParser) isDateValidToFetch(day *time.Time) (bool, error) {
+	if day == nil {
+		return false, fmt.Errorf("invalid day")
+	}
+
+	if o.isMoreThanXDaysInFuture(day, 22) {
+		return false, fmt.Errorf("maximum for days in future for ORF is 22")
+	}
+
+	if o.isMoreThanXDaysInPast(day, 15) {
+		return false, fmt.Errorf("maximum for days in past for ORF is 15")
+	}
+	return true, nil
 }
 
 // helper method to get a collector instance