+++ srf.go
@@ -20,7 +20,6 @@ package main
 import (
 	"fmt"
 	"github.com/PuerkitoBio/goquery"
-	"github.com/alitto/pond"
 	"github.com/gocolly/colly/v2"
 	"log"
 	"math"
@@ -42,72 +41,19 @@ var (
 
 // SRFParser struct for srf parsing code
 type SRFParser struct {
+	ParserInterface
 	Parser
 }
 
-// Fetch central method to parse SRF tv show and program data
-func (s *SRFParser) Fetch() {
-	db := getDb()
-	s.db = db
+func (s *SRFParser) postProcess() {}
 
-	// get channel family db record
-	var channelFamily = getChannelFamily(db, "SRF")
-	if channelFamily.ID == 0 {
-		log.Fatalln("SRF channelFamily was not found!")
-		return
-	}
-	s.ChannelFamily = *channelFamily
-
-	// import tv shows
-	if GetAppConf().EnableTVShowCollection {
-		s.fetchTvShows()
-	}
-
-	// import program entries for the configured date range
-	if GetAppConf().EnableProgramEntryCollection {
-		daysInPast := GetAppConf().DaysInPast
-
-		if daysInPast > 15 {
-			warnMsg := "Maximum for days in past for SRF is 15!\n"
-			log.Printf(warnMsg)
-			appLog(warnMsg)
-			daysInPast = 15
-		}
-
-		timeRange := s.dateRangeHandler.getDateRange()
-		var times []time.Time
-		if timeRange != nil {
-			times = *timeRange
-		} else {
-			log.Printf("No valid time range received!\n")
-			return
-		}
-
-		pool := pond.New(4, 100, getWorkerPoolIdleTimeout())
-		for _, channel := range getChannelsOfFamily(db, channelFamily) {
-			for _, day := range times {
-				if int(time.Since(day).Hours()/24) <= 30 {
-					chn := channel
-					dayToFetch := day
-
-					// srf specific limits!
-					pool.Submit(func() {
-						s.handleDay(chn, dayToFetch)
-					})
-				}
-			}
-		}
-		// wait for finish
-		pool.StopAndWait()
-	}
-
-	if verboseGlobal {
-		log.Println("SRF parsed successfully")
-	}
+func (s *SRFParser) preProcess() bool {
+	s.parallelWorkersCount = 4
+	return true
 }
 
 // fetchTvShowSRF: This method checks all the tv shows
-func (s *SRFParser) fetchTvShows() {
+func (s *SRFParser) fetchTVShows() {
 	if !GetAppConf().EnableTVShowCollection || isRecentlyFetched() {
 		log.Printf("Skip update of tv shows, due to recent fetch. Use 'forceUpdate' = true to ignore this.")
 		return
@@ -149,7 +95,7 @@ func (s *SRFParser) fetchTvShows() {
 
 	err := c.Visit("https://www.srf.ch/play/tv/sendungen")
 	if err != nil {
-		appLog(fmt.Sprintf("Problem fetching URL 'https://www.srf.ch/play/tv/sendungen' %v.\n", err))
+		appLog(fmt.Sprintf("Problem fetching URL 'https://www.srf.ch/play/tv/sendungen' %v.", err))
 	}
 	c.Wait()
 }
@@ -339,9 +285,24 @@ func (s *SRFParser) handleDay(channel Channel, day time.Time) {
 
 	err := c.Visit(queryURL)
 	if err != nil {
-		appLog(fmt.Sprintf("Error of collector: %v\n", err))
+		appLog(fmt.Sprintf("Error of collector: %v", err))
 	}
 	c.Wait()
+}
+
+func (s *SRFParser) isDateValidToFetch(day *time.Time) (bool, error) {
+	if day == nil {
+		return false, fmt.Errorf("invalid day")
+	}
+
+	if s.isMoreThanXDaysInFuture(day, 30) {
+		return false, fmt.Errorf("maximum for days in future for SRF is 30")
+	}
+
+	if s.isMoreThanXDaysInPast(day, 15) {
+		return false, fmt.Errorf("maximum for days in past for SRF is 15")
+	}
+	return true, nil
 }
 
 // getDateFromStringSrf: Extract a date object of a given string format of the page