+++ zdf.go
@@ -46,66 +46,27 @@ var (
 
 // ZDFParser struct to group zdf parsing code
 type ZDFParser struct {
+	ParserInterface
 	Parser
 	zdfAPIKey string
 }
 
-// Fetch This method handles the whole ZDF stuff
-func (z *ZDFParser) Fetch() {
+func (z *ZDFParser) postProcess() {}
+
+func (z *ZDFParser) preProcess() bool {
 	var apiKeyErr error
 	zdfAPIKey, apiKeyErr := z.getZdfAPIKey()
-	if z.zdfAPIKey == "" || apiKeyErr != nil {
+	if apiKeyErr != nil {
 		log.Printf("Error fetching zdf api key: %v\n", apiKeyErr)
-		return
+		return false
 	}
 	if verboseGlobal {
-		log.Println("Start parsing ZDF")
 		log.Printf("Using ZDF API key: %s\n", *zdfAPIKey)
 	}
 	z.zdfAPIKey = *zdfAPIKey
 
-	db := getDb()
-	z.db = db
-	// get channel family db record
-	var channelFamily = getChannelFamily(db, z.ChannelFamilyKey)
-	if channelFamily.ID == 0 {
-		log.Fatalln("ZDF channelFamily was not found! Cancel execution")
-		return
-	}
-	z.ChannelFamily = *channelFamily
-
-	timeRange := z.dateRangeHandler.getDateRange()
-	var times []time.Time
-	if timeRange != nil {
-		times = *timeRange
-	} else {
-		log.Printf("No valid time range received!\n")
-		return
-	}
-	if GetAppConf().EnableTVShowCollection {
-		z.fetchTvShows()
-	}
-
-	// import program entries for the configured date range
-	if GetAppConf().EnableProgramEntryCollection {
-		pool := pond.New(4, 100, getWorkerPoolIdleTimeout())
-		for _, channel := range getChannelsOfFamily(db, channelFamily) {
-			for _, day := range times {
-				chn := channel
-				dayToFetch := day
-
-				pool.Submit(func() {
-					z.handleDayZDF(chn, dayToFetch)
-				})
-			}
-		}
-		// wait for finish
-		pool.StopAndWait()
-	}
-
-	if verboseGlobal {
-		log.Println("ZDF parsed successfully")
-	}
+	z.parallelWorkersCount = 4
+	return true
 }
 
 // getZdfAPIKey method to retrieve the api key we need to connect to the zdf api
@@ -134,7 +95,7 @@ func (z *ZDFParser) getZdfAPIKey() (*string, error) {
 }
 
 // method to process a single day of a single channel
-func (z *ZDFParser) handleDayZDF(channel Channel, day time.Time) {
+func (z *ZDFParser) handleDay(channel Channel, day time.Time) {
 	startDateStr := day.Format(time.RFC3339)
 	endDate := day.AddDate(0, 0, 1)
 	endDateStr := endDate.Format(time.RFC3339)
@@ -325,7 +286,7 @@ func (z *ZDFParser) doZDFApiProgramItemRequest(apiURL string) (*ZdfProgramItemRe
 }
 
 // fetchTvShows method to fetch zdf tv shows
-func (z *ZDFParser) fetchTvShows() {
+func (z *ZDFParser) fetchTVShows() {
 	if !GetAppConf().EnableTVShowCollection || isRecentlyFetched() {
 		log.Printf("Skip update of tv shows, due to recent fetch. Use 'forceUpdate' = true to ignore this.")
 		return
@@ -342,7 +303,7 @@ func (z *ZDFParser) fetchTvShows() {
 
 		document, err := getDocument(apiURL)
 		if document == nil || err != nil {
-			appLog(fmt.Sprintf("Problem with http call to zdf %v\n", err))
+			appLog(fmt.Sprintf("Problem with http call to zdf %v", err))
 			continue
 		}
 		document.Find(".b-content-teaser-item h3 a").Each(func(i int, selection *goquery.Selection) {
@@ -410,6 +371,22 @@ func (z *ZDFParser) processSingleTvShow(singleTvShowPage string) {
 
 	tvShowRecord.saveTvShowRecord(z.db)
 	return
+}
+
+func (z *ZDFParser) isDateValidToFetch(day *time.Time) (bool, error) {
+	if day == nil {
+		return false, fmt.Errorf("invalid day")
+	}
+
+	if z.isMoreThanXDaysInFuture(day, 43) { // = six weeks in future + today
+		return false, fmt.Errorf("maximum for days in future for ZDF is 43")
+	}
+	location, _ := time.LoadLocation(GetAppConf().TimeZone)
+	earliestDate := time.Date(2011, 1, 1, 0, 0, 0, 0, location)
+	if day.Before(earliestDate) {
+		return false, fmt.Errorf("maximum for days in past for ZDF is %s", earliestDate.Format(time.RFC822))
+	}
+	return true, nil
 }
 
 // ZdfBroadcastResponse api response struct definitions